compatibility
func add changed to int64 so website affected.

+++ cmd/ogamed/myfunc.go
@@ -66,7 +66,7 @@ func RangeStructer(args ...interface{}) []interface{} {
 	return out
 }
 
-func add(x int, y int) int {
+func add(x int64, y int64) int64 {
 	return x + y
 }
